moved the ui files in main to src/ui
