+++ src/main.py
@@ -1,11 +1,5 @@
 #!/usr/bin/env python3
 
-import os
-import sys
-import subprocess
-
-from PyQt6.QtWidgets import (QMainWindow,QApplication,QVBoxLayout,QWidget)
-
 
 #from src.utils import app_lister
 from utils import app_lister
@@ -14,22 +8,3 @@ from utils import app_lister
 apps = app_lister.application_lister()
 
 
-class MainWindow(QMainWindow):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args,**kwargs)
-
-        container = QWidget()
-        containerLayout = QVBoxLayout()
-        container.setLayout(containerLayout)
-
-        self.setCentralWidget(container)
-
-
-
-
-
-app = QApplication(sys.argv)
-w = MainWindow()
-w.show()
-sys.exit(app.exec())
